+++ src/extension.js
@@ -4,8 +4,6 @@ const vscode = require('vscode');
 const { runner } = require('./language/general/runner');
 const { runner: javascriptRunner } = require('./language/javascript/runner');
 const { runner: phpRunner } = require('./language/php/runner');
-const { resetCache } = require('./language/general/providers');
-const HintList = require('./language/general/hintList');
 
 const hintDecorationType = vscode.window.createTextEditorDecorationType({});
 // this method is called when your extension is activated
@@ -29,18 +27,12 @@ function activate(context) {
 	);
 
 	let timeout = null;
-	const trigger = (identifier, editor, force, time = 100, reset = false) => {
+	const trigger = (identifier, editor, force, time = 100) => {
 		if (currentRunner && !currentRunner.state.done) {
 			currentRunner.reject();
-			console.log('stopped previous runner');
 		}
-		console.log(identifier, time);
 		if (timeout) {
 			clearTimeout(timeout);
-		}
-		if (reset) {
-			HintList.clear();
-			resetCache();
 		}
 		timeout = setTimeout(() => {
 			if (editor && (isEnabled() || force)) {
@@ -82,17 +74,22 @@ function activate(context) {
 		vscode.window.setStatusBarMessage(message, hideMessageAfterMs);
 	})
 
-	trigger('on start', activeEditor, false, 100, true);
+	trigger('on start', activeEditor, false, 100);
 
 	context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor(editor => {
 		activeEditor = editor;
-		trigger('change_active_text_editor', activeEditor, false, 100, true);
+		trigger('change_active_text_editor', activeEditor, false, 100);
 	}));
 
 	context.subscriptions.push(vscode.workspace.onDidChangeTextDocument(event => {
 		if (event.contentChanges.length) {
 			trigger('text edited', activeEditor, false, 300);
 		}
+	}))
+
+	context.subscriptions.push(vscode.window.onDidChangeTextEditorVisibleRanges(event => {
+		activeEditor = event.textEditor;
+		trigger('scroll', activeEditor, false, 100);
 	}))
 }
 exports.activate = activate;